:art: [#102] refactor: Unify imports and apply __future__ annotations
In an attempt to find the source of the str vs Struct warning I unified
all the import paths and used `from __future__ import annotations`,
because that makes all annotations string-like.

Leaving it in for future proofing, better to bump into issues early than
have an massive refactor when this lands in standard cpython.
