+++ backend/src/openbeheer/types/__init__.py
@@ -1,29 +1,34 @@
 """Data Transfer Object types for both ZGW APIs facing side, and our frontend."""
 
 from ._open_beheer import (
+    DetailResponse,
     ExternalServiceError,
+    FrontendFieldSet,
+    FrontendFieldsets,
     OBField,
     OBFieldType,
     OBList,
     OBOption,
     OBPagedQueryParams,
     OBPagination,
+    VersionSummary,
     as_ob_fieldtype,
     as_ob_option,
 )
 from ._zgw import ZGWError, ZGWResponse
-from .ztc import Catalogus, StatusType
 
 __all__ = [
-    "Catalogus",
+    "DetailResponse",
     "ExternalServiceError",
+    "FrontendFieldSet",
+    "FrontendFieldsets",
     "OBField",
     "OBFieldType",
     "OBList",
     "OBOption",
-    "OBPagination",
     "OBPagedQueryParams",
-    "StatusType",
+    "OBPagination",
+    "VersionSummary",
     "ZGWError",
     "ZGWResponse",
     "as_ob_fieldtype",
